Only unlink the file if there is clearance, else, since the file is still in temporary storage, it'll be removed by php.

+++ src/Uploader/Uploader.php
@@ -123,11 +123,12 @@ class Uploader
                 'error_message' => 'Invalid Credentials',
             ];
         }
-
-        if (null != $file_name) {
-            unlink(getenv('TMP_STORE').$file_name);
-        } else {
-            unlink(implode('', $file['tmp_name']));
+        if (true == $check_against_hashlist['clearance']) {
+            if (null != $file_name) {
+                unlink(getenv('TMP_STORE').$file_name);
+            } else {
+                unlink(implode('', $file['tmp_name']));
+            }
         }
 
         return $response;
